Fix compilation on GCC/Clang with missing header

+++ codehero/ui/fontface_freetype.cpp
@@ -7,6 +7,7 @@
 #include "core/image.h"
 #include "core/enginecontext.h"
 #include "graphics/rendersystem.h"
+#include <cmath>
 #include <cstdint>
 #include <ft2build.h>
 #include FT_FREETYPE_H
@@ -41,8 +42,8 @@ void FontFaceFreeType::_Load() {
     }
 
     {
-        m_Ascender = std::floorf((face->size->metrics.ascender >> 6) + 0.5f); // / 64
-        m_MaxHeight = std::floorf(((face->size->metrics.ascender - face->size->metrics.descender) >> 6) + 0.5f);
+        m_Ascender = std::floor((face->size->metrics.ascender >> 6) + 0.5f); // / 64
+        m_MaxHeight = std::floor(((face->size->metrics.ascender - face->size->metrics.descender) >> 6) + 0.5f);
         // List all glyph
         uint32_t numGlyphs = static_cast<uint32_t>(face->num_glyphs);
         std::vector<uint32_t> charCodes(numGlyphs, 0);
